scripts/requirements.txt
# YCharts API Query Script Requirements
git+https://github.com/ycharts/pycharts.git
requests>=2.25.0
pandas>=1.3.0
python-dotenv>=0.19.0
